+++ src/SWP/Bundle/ContentBundle/Model/Article.php
@@ -323,4 +323,16 @@ class Article implements ArticleInterface, MediaAwareArticleInterface
     {
         $this->templateName = $templateName;
     }
+
+    /**
+     * Don't serialize values.
+     *
+     * @return array
+     */
+    public function __sleep()
+    {
+        $this->media = "Cannot be serializable";
+
+        return array_keys(get_object_vars($this));
+    }
 }